+++ assets/libft/src/ft_substr.c
@@ -6,16 +6,16 @@
 /*   By: dbredykh <dbredykh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/23 16:42:07 by dbredykh          #+#    #+#             */
-/*   Updated: 2023/06/19 15:59:16 by dbredykh         ###   ########.fr       */
+/*   Updated: 2024/01/08 23:11:00 by dbredykh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*dest;
-	size_t	i;
+	int		i;
 
 	if (start >= ft_strlen(s))
 		return (ft_strdup(""));
